make int literals work as floats

+++ tests/tests.rs
@@ -3,14 +3,14 @@ use serde::Deserialize;
 
 #[test]
 fn to_serde() {
-    #[derive(Debug, PartialEq, Eq, Deserialize)]
+    #[derive(Debug, PartialEq, Deserialize)]
     struct Target {
         an_enum: AnEnum,
         a_kid: Option<Kiddo>,
     }
-    #[derive(Debug, PartialEq, Eq, Deserialize)]
-    struct Kiddo(i32, i32, i32);
-    #[derive(Debug, PartialEq, Eq, Deserialize)]
+    #[derive(Debug, PartialEq, Deserialize)]
+    struct Kiddo(i32, u32, f32);
+    #[derive(Debug, PartialEq, Deserialize)]
     enum AnEnum {
         Variant1,
         Variant2(String),
@@ -36,7 +36,7 @@ fn to_serde() {
         vec![
             Target {
                 an_enum: AnEnum::Variant1,
-                a_kid: Some(Kiddo(1, 2, 3))
+                a_kid: Some(Kiddo(1i32, 2u32, 3.0))
             },
             Target {
                 an_enum: AnEnum::Variant2("hello, world".into()),
